added css for correct svg transitions in firefox
