+++ public/locales/uk/common.json
@@ -101,6 +101,7 @@
     "pageNotFound": "Сторінку не знайдено"
   },
   "utils": {
-    "play": "відтворити"
+    "play": "відтворити",
+    "new_game": "Нова гра"
   }
 }
